Reorganized the libnetwork code to seperate Controller, Network and Endpoint
